Move `BUILTIN_FIELDS` to `stateful::agent`
