+++ packages/engine/lib/stateful/src/agent/mod.rs
@@ -1,4 +1,7 @@
 mod field;
 mod schema;
 
-pub use self::{field::AgentStateField, schema::AgentSchema};
+pub use self::{
+    field::{AgentStateField, BUILTIN_FIELDS},
+    schema::AgentSchema,
+};
